add TLS integration test for AddPackageRepository re-organize integration tests by the functionality
