+++ chart/kubeapps/Chart.lock
@@ -1,12 +1,12 @@
 dependencies:
 - name: common
   repository: https://charts.bitnami.com/bitnami
-  version: 1.11.1
+  version: 1.11.2
 - name: postgresql
   repository: https://charts.bitnami.com/bitnami
   version: 10.16.2
 - name: redis
   repository: https://charts.bitnami.com/bitnami
   version: 15.7.6
-digest: sha256:a787a6ab3f33bd0f630c4636fac61264fc0e12d6e0d12604fb318da188fcc8a4
-generated: "2022-02-24T16:42:03.375120864Z"
+digest: sha256:e2fbacc0528c8d10ce61014c4cbaf52186dbcf66d09c1d05fe0a8078f7f6bcda
+generated: "2022-03-01T17:48:06.944807-08:00"
