+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/release_integration_test.go
@@ -5,7 +5,6 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"strings"
 	"testing"
 	"time"
@@ -15,11 +14,8 @@ import (
 	corev1 "github.com/kubeapps/kubeapps/cmd/kubeapps-apis/gen/core/packages/v1alpha1"
 	plugins "github.com/kubeapps/kubeapps/cmd/kubeapps-apis/gen/core/plugins/v1alpha1"
 	fluxplugin "github.com/kubeapps/kubeapps/cmd/kubeapps-apis/gen/plugins/fluxv2/packages/v1alpha1"
-	"golang.org/x/sync/semaphore"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	apiv1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
 )
 
 // This is an integration test: it tests the full integration of flux plugin with flux back-end
@@ -29,18 +25,6 @@ import (
 // 2) kubeapps apis apiserver service running with fluxv2 plug-in enabled, port forwarded to 8080, e.g.
 //      kubectl -n kubeapps port-forward svc/kubeapps-internal-kubeappsapis 8080:8080
 // 3) run './kind-cluster-setup.sh deploy' once prior to these tests
-
-const (
-	// This is local copy of the first few entries
-	// on "https://stefanprodan.github.io/podinfo/index.yaml" as of Sept 10 2021 with the chart
-	// urls modified to link to .tgz files also within the local cluster.
-	// If we want other repos, we'll have add directories and tinker with ./Dockerfile and NGINX conf.
-	// This relies on fluxv2plugin-testdata-svc service stood up by testdata/kind-cluster-setup.sh
-	podinfo_repo_url = "http://fluxv2plugin-testdata-svc.default.svc.cluster.local:80/podinfo"
-
-	// same as above but requires HTTP basic authentication: user: foo, password: bar
-	podinfo_basic_auth_repo_url = "http://fluxv2plugin-testdata-svc.default.svc.cluster.local:80/podinfo-basic-auth"
-)
 
 type integrationTestCreatePackageSpec struct {
 	testName          string
@@ -403,531 +387,6 @@ func TestKindClusterDeleteInstalledPackage(t *testing.T) {
 					t.Logf("Waiting 2s for garbage collection of [%s], attempt [%d/%d]...", pods[0], i+1, maxWait)
 					time.Sleep(2 * time.Second)
 				}
-			}
-		})
-	}
-}
-
-// this integration test is meant to test a scenario when the redis cache is confiured with maxmemory
-// too small to be able to fit all the repos needed to satisfy the request for GetAvailablePackageSummaries
-// and redis cache eviction kicks in. Also, the kubeapps-apis pod should have a large memory limit (1Gb) set
-// To set up such environment one can use  "-f ./docs/user/manifests/kubeapps-local-dev-redis-tiny-values.yaml"
-// option when installing kubeapps via "helm upgrade"
-// It is worth noting that exactly how many copies of bitnami repo can be held in the cache at any given time varies
-// This is because the size of the index.yaml we get from bitnami does fluctuate quite a bit over time:
-// [kubeapps]$ ls -l bitnami_index.yaml
-// -rw-r--r--@ 1 gfichtenholt  staff  8432962 Jun 20 02:35 bitnami_index.yaml
-// [kubeapps]$ ls -l bitnami_index.yaml
-// -rw-rw-rw-@ 1 gfichtenholt  staff  10394218 Nov  7 19:41 bitnami_index.yaml
-// Also now we are caching helmcharts themselves for each repo so that will affect how many will fit too
-func TestKindClusterGetAvailablePackageSummariesForLargeReposAndTinyRedis(t *testing.T) {
-	fluxPlugin, _ := checkEnv(t)
-
-	redisCli, err := newRedisClientForIntegrationTest(t)
-	if err != nil {
-		t.Fatalf("%+v", err)
-	}
-
-	// assume 30Mb redis cache for now. See comment above
-	if err = redisCheckTinyMaxMemory(t, redisCli, "31457280"); err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	// ref https://redis.io/topics/notifications
-	if err = redisCli.ConfigSet(redisCli.Context(), "notify-keyspace-events", "EA").Err(); err != nil {
-		t.Fatalf("%+v", err)
-	}
-	t.Cleanup(func() {
-		t.Logf("Resetting notify-keyspace-events")
-		if err = redisCli.ConfigSet(redisCli.Context(), "notify-keyspace-events", "").Err(); err != nil {
-			t.Logf("%v", err)
-		}
-	})
-
-	if err = initNumberOfChartsInBitnamiCatalog(t); err != nil {
-		t.Errorf("Failed to get number of charts in bitnami catalog due to: %v", err)
-	}
-
-	const MAX_REPOS_NEVER = 100
-	var totalRepos = 0
-	// ref https://stackoverflow.com/questions/32840687/timeout-for-waitgroup-wait
-	evictedRepos := sets.String{}
-
-	// do this part in a func so we can defer subscribe.Close
-	func() {
-		// ref https://medium.com/nerd-for-tech/redis-getting-notified-when-a-key-is-expired-or-changed-ca3e1f1c7f0a
-		subscribe := redisCli.PSubscribe(redisCli.Context(), "__keyevent@0__:*")
-		defer subscribe.Close()
-
-		sem := semaphore.NewWeighted(MAX_REPOS_NEVER)
-		if err := sem.Acquire(context.Background(), MAX_REPOS_NEVER); err != nil {
-			t.Fatalf("%v", err)
-		}
-
-		go redisReceiveNotificationsLoop(t, subscribe.Channel(), sem, &evictedRepos)
-
-		// now load some large repos (bitnami)
-		// I didn't want to store a large (>10MB) copy of bitnami repo in our git,
-		// so for now let it fetch directly from bitnami website
-		// we'll keep adding repos one at a time, until we get an event from redis
-		// about the first evicted repo entry
-		for ; totalRepos < MAX_REPOS_NEVER && evictedRepos.Len() == 0; totalRepos++ {
-			repo := fmt.Sprintf("bitnami-%d", totalRepos)
-			// this is to make sure we allow enough time for repository to be created and come to ready state
-			if err = kubeAddHelmRepository(t, repo, "https://charts.bitnami.com/bitnami", "default", ""); err != nil {
-				t.Fatalf("%v", err)
-			}
-			t.Cleanup(func() {
-				if err = kubeDeleteHelmRepository(t, repo, "default"); err != nil {
-					t.Logf("%v", err)
-				}
-			})
-			// wait until this repo have been indexed and cached up to 10 minutes
-			ctx, cancel := context.WithTimeout(context.Background(), time.Minute*10)
-			defer cancel()
-			if err := sem.Acquire(ctx, 1); err != nil {
-				t.Fatalf("Timed out waiting for Redis event: %v", err)
-			}
-		}
-		t.Logf("Done with first part of the test, total repos: [%d], evicted repos: [%d]",
-			totalRepos, len(evictedRepos))
-	}()
-
-	if evictedRepos.Len() == 0 {
-		t.Fatalf("Failing because redis did not evict any entries")
-	}
-
-	if keys, err := redisCli.Keys(redisCli.Context(), "helmrepositories:*").Result(); err != nil {
-		t.Fatalf("%v", err)
-	} else {
-		// the cache should only big enough to be able to hold at most (totalRepos-1) of the keys
-		// one (or more) entries may have been evicted
-		if len(keys) > totalRepos-1 {
-			t.Fatalf("Expected at most [%d] keys in cache but got: %s", totalRepos-1, keys)
-		}
-	}
-
-	// one particular code path I'd like to test:
-	// make sure that GetAvailablePackageVersions() works w.r.t. a cache entry that's been evicted
-	grpcContext := newGrpcAdminContext(t, "test-create-admin")
-
-	// copy the evicted list because before ForEach loop below will modify it in a goroutine
-	evictedCopy := sets.StringKeySet(evictedRepos)
-
-	// do this part in a func so we can defer subscribe.Close
-	func() {
-		subscribe := redisCli.PSubscribe(redisCli.Context(), "__keyevent@0__:*")
-		defer subscribe.Close()
-
-		go redisReceiveNotificationsLoop(t, subscribe.Channel(), nil, &evictedRepos)
-
-		for _, k := range evictedCopy.List() {
-			name := strings.Split(k, ":")[2]
-			t.Logf("Checking apache version in repo [%s]...", name)
-			grpcContext, cancel := context.WithTimeout(grpcContext, defaultContextTimeout)
-			defer cancel()
-			resp, err := fluxPlugin.GetAvailablePackageVersions(
-				grpcContext, &corev1.GetAvailablePackageVersionsRequest{
-					AvailablePackageRef: &corev1.AvailablePackageReference{
-						Context: &corev1.Context{
-							Namespace: "default",
-						},
-						Identifier: name + "/apache",
-					},
-				})
-			if err != nil {
-				t.Fatalf("%v", err)
-			} else if len(resp.PackageAppVersions) < 5 {
-				t.Fatalf("Expected at least 5 versions for apache chart, got: %s", resp)
-			}
-		}
-
-		t.Logf("Done with second part of the test")
-	}()
-
-	// do this part in a func so we can defer subscribe.Close
-	func() {
-		subscribe := redisCli.PSubscribe(redisCli.Context(), "__keyevent@0__:*")
-		defer subscribe.Close()
-
-		// above loop should cause a few more entries to be evicted, but just to be sure let's
-		// load a few more copies of bitnami repo into the cache. The goal of this for loop is
-		// to force redis to evict more repo(s)
-		sem := semaphore.NewWeighted(MAX_REPOS_NEVER)
-		if err := sem.Acquire(context.Background(), MAX_REPOS_NEVER); err != nil {
-			t.Fatalf("%v", err)
-		}
-		go redisReceiveNotificationsLoop(t, subscribe.Channel(), sem, &evictedRepos)
-
-		for ; totalRepos < MAX_REPOS_NEVER && evictedRepos.Len() == evictedCopy.Len(); totalRepos++ {
-			repo := fmt.Sprintf("bitnami-%d", totalRepos)
-			// this is to make sure we allow enough time for repository to be created and come to ready state
-			if err = kubeAddHelmRepository(t, repo, "https://charts.bitnami.com/bitnami", "default", ""); err != nil {
-				t.Fatalf("%v", err)
-			}
-			t.Cleanup(func() {
-				if err = kubeDeleteHelmRepository(t, repo, "default"); err != nil {
-					t.Logf("%v", err)
-				}
-			})
-			// wait until this repo have been indexed and cached up to 10 minutes
-			ctx, cancel := context.WithTimeout(context.Background(), time.Minute*10)
-			defer cancel()
-			if err := sem.Acquire(ctx, 1); err != nil {
-				t.Fatalf("Timed out waiting for Redis event: %v", err)
-			}
-		}
-
-		t.Logf("Done with third part of the test")
-	}()
-
-	if keys, err := redisCli.Keys(redisCli.Context(), "helmrepositories:*").Result(); err != nil {
-		t.Fatalf("%v", err)
-	} else {
-		// the cache should only big enough to be able to hold at most (totalRepos-1) of the keys
-		// one (or more) entries MUST have been evicted
-		if len(keys) > totalRepos-1 {
-			t.Fatalf("Expected at most %d keys in cache but got [%s]", totalRepos-1, keys)
-		}
-	}
-
-	// not related to low maxmemory but as long as we are here might as well check that
-	// there is a Unauthenticated failure when there are no credenitals in the request
-	_, err = fluxPlugin.GetAvailablePackageSummaries(context.TODO(), &corev1.GetAvailablePackageSummariesRequest{})
-	if err == nil || status.Code(err) != codes.Unauthenticated {
-		t.Fatalf("Expected Unauthenticated, got %v", err)
-	}
-
-	grpcContext, cancel := context.WithTimeout(grpcContext, 60*time.Second)
-	defer cancel()
-	resp2, err := fluxPlugin.GetAvailablePackageSummaries(grpcContext, &corev1.GetAvailablePackageSummariesRequest{})
-	if err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	// we need to make sure that response contains packages from all existing repositories
-	// regardless whether they're in the cache or not
-	expected := sets.String{}
-	for i := 0; i < totalRepos; i++ {
-		repo := fmt.Sprintf("bitnami-%d", i)
-		expected.Insert(repo)
-	}
-	for _, s := range resp2.AvailablePackageSummaries {
-		id := strings.Split(s.AvailablePackageRef.Identifier, "/")
-		expected.Delete(id[0])
-	}
-
-	if expected.Len() != 0 {
-		t.Fatalf("Expected to get packages from these repositories: %s, but did not get any",
-			expected.List())
-	}
-}
-
-// this test is testing a scenario when a repo that takes a long time to index is added
-// and while the indexing is in progress this repo is deleted by another request.
-// The goal is to make sure that the events are processed by the cache fully in the order
-// they were received and the cache does not end up in inconsistent state
-func TestKindClusterAddThenDeleteRepo(t *testing.T) {
-	checkEnv(t)
-
-	redisCli, err := newRedisClientForIntegrationTest(t)
-	if err != nil {
-		t.Fatalf("%+v", err)
-	}
-
-	// now load some large repos (bitnami)
-	// I didn't want to store a large (10MB) copy of bitnami repo in our git,
-	// so for now let it fetch from bitnami website
-	if err = kubeAddHelmRepository(t, "bitnami-1", "https://charts.bitnami.com/bitnami", "default", ""); err != nil {
-		t.Fatalf("%v", err)
-	}
-	// wait until this repo reaches 'Ready' state so that long indexation process kicks in
-	if err = kubeWaitUntilHelmRepositoryIsReady(t, "bitnami-1", "default"); err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	if err = kubeDeleteHelmRepository(t, "bitnami-1", "default"); err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	t.Logf("Waiting up to 30 seconds...")
-	time.Sleep(30 * time.Second)
-
-	if keys, err := redisCli.Keys(redisCli.Context(), "*").Result(); err != nil {
-		t.Fatalf("%v", err)
-	} else {
-		if len(keys) != 0 {
-			t.Fatalf("Failing due to unexpected state of the cache. Current keys: %s", keys)
-		}
-	}
-}
-
-func TestKindClusterRepoWithBasicAuth(t *testing.T) {
-	fluxPluginClient, _ := checkEnv(t)
-
-	secretName := "podinfo-basic-auth-secret"
-	repoName := "podinfo-basic-auth"
-
-	if err := kubeCreateSecret(t, newBasicAuthSecret(secretName, "default", "foo", "bar")); err != nil {
-		t.Fatalf("%v", err)
-	}
-	t.Cleanup(func() {
-		err := kubeDeleteSecret(t, "default", secretName)
-		if err != nil {
-			t.Logf("Failed to delete helm repository due to [%v]", err)
-		}
-	})
-
-	if err := kubeAddHelmRepository(t, repoName, podinfo_basic_auth_repo_url, "default", secretName); err != nil {
-		t.Fatalf("%v", err)
-	}
-	t.Cleanup(func() {
-		err := kubeDeleteHelmRepository(t, repoName, "default")
-		if err != nil {
-			t.Logf("Failed to delete helm repository due to [%v]", err)
-		}
-	})
-
-	// wait until this repo reaches 'Ready'
-	if err := kubeWaitUntilHelmRepositoryIsReady(t, repoName, "default"); err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	grpcContext := newGrpcAdminContext(t, "test-create-admin-basic-auth")
-
-	const maxWait = 25
-	for i := 0; i <= maxWait; i++ {
-		grpcContext, cancel := context.WithTimeout(grpcContext, defaultContextTimeout)
-		defer cancel()
-		resp, err := fluxPluginClient.GetAvailablePackageSummaries(
-			grpcContext,
-			&corev1.GetAvailablePackageSummariesRequest{
-				Context: &corev1.Context{
-					Namespace: "default",
-				},
-			})
-		if err == nil {
-			opt1 := cmpopts.IgnoreUnexported(
-				corev1.GetAvailablePackageSummariesResponse{},
-				corev1.AvailablePackageSummary{},
-				corev1.AvailablePackageReference{},
-				corev1.Context{},
-				plugins.Plugin{},
-				corev1.PackageAppVersion{})
-			opt2 := cmpopts.SortSlices(lessAvailablePackageFunc)
-			if got, want := resp, available_package_summaries_podinfo_basic_auth; !cmp.Equal(got, want, opt1, opt2) {
-				t.Errorf("mismatch (-want +got):\n%s", cmp.Diff(want, got, opt1, opt2))
-			}
-			break
-		} else if i == maxWait {
-			t.Fatalf("Timed out waiting for available package summaries, last response: %v, last error: [%v]", resp, err)
-		} else {
-			t.Logf("Waiting 2s for repository [%s] to be indexed, attempt [%d/%d]...", repoName, i+1, maxWait)
-			time.Sleep(2 * time.Second)
-		}
-	}
-
-	availablePackageRef := availableRef(repoName+"/podinfo", "default")
-
-	// first try the negative case, no auth - should fail due to not being able to
-	// read secrets in all namespaces
-	fluxPluginServiceAccount := "test-repo-with-basic-auth"
-	ctx, cancel := context.WithTimeout(newGrpcFluxPluginContext(t, fluxPluginServiceAccount), defaultContextTimeout)
-	defer cancel()
-	_, err := fluxPluginClient.GetAvailablePackageDetail(
-		ctx,
-		&corev1.GetAvailablePackageDetailRequest{AvailablePackageRef: availablePackageRef})
-	if err == nil {
-		t.Fatalf("Expected error, did not get one")
-	} else if status.Code(err) != codes.PermissionDenied {
-		t.Fatalf("GetAvailablePackageDetailRequest expected: PermissionDenied, got: %v", err)
-	}
-
-	// this should succeed as it is done in the context of cluster admin
-	grpcContext, cancel = context.WithTimeout(grpcContext, defaultContextTimeout)
-	defer cancel()
-	resp, err := fluxPluginClient.GetAvailablePackageDetail(
-		grpcContext,
-		&corev1.GetAvailablePackageDetailRequest{AvailablePackageRef: availablePackageRef})
-	if err != nil {
-		t.Fatalf("%v", err)
-	}
-
-	compareActualVsExpectedAvailablePackageDetail(t, resp.AvailablePackageDetail, expected_detail_podinfo_basic_auth.AvailablePackageDetail)
-}
-
-type integrationTestAddRepoSpec struct {
-	testName                 string
-	request                  *corev1.AddPackageRepositoryRequest
-	existingSecret           *apiv1.Secret
-	expectedResponse         *corev1.AddPackageRepositoryResponse
-	expectedStatusCode       codes.Code
-	expectedReconcileFailure bool
-}
-
-func TestKindClusterAddPackageRepository(t *testing.T) {
-	_, fluxPluginReposClient := checkEnv(t)
-
-	testCases := []integrationTestAddRepoSpec{
-		{
-			testName: "add repo test (simplest case)",
-			request: &corev1.AddPackageRepositoryRequest{
-				Name:    "my-podinfo",
-				Context: &corev1.Context{Namespace: "default"},
-				Type:    "helm",
-				Url:     podinfo_repo_url,
-			},
-			expectedResponse: &corev1.AddPackageRepositoryResponse{
-				PackageRepoRef: &corev1.PackageRepositoryReference{
-					Context: &corev1.Context{
-						Namespace: "default",
-						Cluster:   KubeappsCluster,
-					},
-					Identifier: "my-podinfo",
-					Plugin:     fluxPlugin,
-				},
-			},
-			expectedStatusCode: codes.OK,
-		},
-		{
-			testName: "package repository with basic auth",
-			request: &corev1.AddPackageRepositoryRequest{
-				Name:    "my-podinfo-2",
-				Context: &corev1.Context{Namespace: "default"},
-				Type:    "helm",
-				Url:     podinfo_basic_auth_repo_url,
-				Auth: &corev1.PackageRepositoryAuth{
-					Type: corev1.PackageRepositoryAuth_PACKAGE_REPOSITORY_AUTH_TYPE_BASIC_AUTH,
-					PackageRepoAuthOneOf: &corev1.PackageRepositoryAuth_UsernamePassword{
-						UsernamePassword: &corev1.UsernamePassword{
-							Username: "foo",
-							Password: "bar",
-						},
-					},
-				},
-			},
-			expectedResponse: &corev1.AddPackageRepositoryResponse{
-				PackageRepoRef: &corev1.PackageRepositoryReference{
-					Context: &corev1.Context{
-						Namespace: "default",
-						Cluster:   KubeappsCluster,
-					},
-					Identifier: "my-podinfo-2",
-					Plugin:     fluxPlugin,
-				},
-			},
-			expectedStatusCode: codes.OK,
-		},
-		{
-			testName: "package repository with wrong basic auth fails",
-			request: &corev1.AddPackageRepositoryRequest{
-				Name:    "my-podinfo-3",
-				Context: &corev1.Context{Namespace: "default"},
-				Type:    "helm",
-				Url:     podinfo_basic_auth_repo_url,
-				Auth: &corev1.PackageRepositoryAuth{
-					Type: corev1.PackageRepositoryAuth_PACKAGE_REPOSITORY_AUTH_TYPE_BASIC_AUTH,
-					PackageRepoAuthOneOf: &corev1.PackageRepositoryAuth_UsernamePassword{
-						UsernamePassword: &corev1.UsernamePassword{
-							Username: "foo",
-							Password: "bar-2",
-						},
-					},
-				},
-			},
-			expectedResponse: &corev1.AddPackageRepositoryResponse{
-				PackageRepoRef: &corev1.PackageRepositoryReference{
-					Context: &corev1.Context{
-						Namespace: "default",
-						Cluster:   KubeappsCluster,
-					},
-					Identifier: "my-podinfo-3",
-					Plugin:     fluxPlugin,
-				},
-			},
-			expectedStatusCode:       codes.OK,
-			expectedReconcileFailure: true,
-		},
-		{
-			testName: "package repository with basic auth and existing secret",
-			request: &corev1.AddPackageRepositoryRequest{
-				Name:    "my-podinfo-4",
-				Context: &corev1.Context{Namespace: "default"},
-				Type:    "helm",
-				Url:     podinfo_basic_auth_repo_url,
-				Auth: &corev1.PackageRepositoryAuth{
-					Type: corev1.PackageRepositoryAuth_PACKAGE_REPOSITORY_AUTH_TYPE_BASIC_AUTH,
-					PackageRepoAuthOneOf: &corev1.PackageRepositoryAuth_SecretRef{
-						SecretRef: &corev1.SecretKeyReference{
-							Name: "secret-1",
-						},
-					},
-				},
-			},
-			existingSecret: newBasicAuthSecret("secret-1", "default", "foo", "bar"),
-			expectedResponse: &corev1.AddPackageRepositoryResponse{
-				PackageRepoRef: &corev1.PackageRepositoryReference{
-					Context: &corev1.Context{
-						Namespace: "default",
-						Cluster:   KubeappsCluster,
-					},
-					Identifier: "my-podinfo-4",
-					Plugin:     fluxPlugin,
-				},
-			},
-			expectedStatusCode: codes.OK,
-		},
-	}
-
-	grpcContext := newGrpcAdminContext(t, "test-add-repo-admin")
-
-	for _, tc := range testCases {
-		t.Run(tc.testName, func(t *testing.T) {
-			ctx, cancel := context.WithTimeout(grpcContext, defaultContextTimeout)
-			defer cancel()
-			if tc.existingSecret != nil {
-				if err := kubeCreateSecret(t, tc.existingSecret); err != nil {
-					t.Fatalf("%v", err)
-				}
-				t.Cleanup(func() {
-					err := kubeDeleteSecret(t, tc.existingSecret.Namespace, tc.existingSecret.Name)
-					if err != nil {
-						t.Logf("Failed to delete secret due to [%v]", err)
-					}
-				})
-			}
-			resp, err := fluxPluginReposClient.AddPackageRepository(ctx, tc.request)
-			if tc.expectedStatusCode != codes.OK {
-				if status.Code(err) != tc.expectedStatusCode {
-					t.Fatalf("Expected %v, got: %v", tc.expectedStatusCode, err)
-				}
-				return // done, nothing more to check
-			} else if err != nil {
-				t.Fatal(err)
-			}
-			t.Cleanup(func() {
-				err := kubeDeleteHelmRepository(t, tc.request.Name, tc.request.Context.Namespace)
-				if err != nil {
-					t.Logf("Failed to delete helm source due to [%v]", err)
-				}
-			})
-			opt1 := cmpopts.IgnoreUnexported(
-				corev1.AddPackageRepositoryResponse{},
-				corev1.Context{},
-				corev1.PackageRepositoryReference{},
-				plugins.Plugin{},
-			)
-			if got, want := resp, tc.expectedResponse; !cmp.Equal(got, want, opt1) {
-				t.Errorf("mismatch (-want +got):\n%s", cmp.Diff(want, got, opt1))
-			}
-
-			// TODO wait for reconcile. To do it properly, we need "R" in CRUD to be
-			// designed and implemented
-			err = kubeWaitUntilHelmRepositoryIsReady(t, tc.request.Name, tc.request.Context.Namespace)
-			if err != nil && !tc.expectedReconcileFailure {
-				t.Fatal(err)
-			} else if err == nil && tc.expectedReconcileFailure {
-				t.Fatalf("Expected error got nil")
 			}
 		})
 	}
@@ -1791,37 +1250,6 @@ var (
 		TargetContext: &corev1.Context{
 			Namespace: "test-15",
 			Cluster:   KubeappsCluster,
-		},
-	}
-
-	available_package_summaries_podinfo_basic_auth = &corev1.GetAvailablePackageSummariesResponse{
-		AvailablePackageSummaries: []*corev1.AvailablePackageSummary{
-			{
-				Name:                "podinfo",
-				AvailablePackageRef: availableRef("podinfo-basic-auth/podinfo", "default"),
-				LatestVersion:       &corev1.PackageAppVersion{PkgVersion: "6.0.0", AppVersion: "6.0.0"},
-				DisplayName:         "podinfo",
-				ShortDescription:    "Podinfo Helm chart for Kubernetes",
-				Categories:          []string{""},
-			},
-		},
-	}
-
-	expected_detail_podinfo_basic_auth = &corev1.GetAvailablePackageDetailResponse{
-		AvailablePackageDetail: &corev1.AvailablePackageDetail{
-			AvailablePackageRef: availableRef("podinfo-basic-auth/podinfo", "default"),
-			Name:                "podinfo",
-			Version:             &corev1.PackageAppVersion{PkgVersion: "6.0.0", AppVersion: "6.0.0"},
-			RepoUrl:             "http://fluxv2plugin-testdata-svc.default.svc.cluster.local:80/podinfo-basic-auth",
-			HomeUrl:             "https://github.com/stefanprodan/podinfo",
-			DisplayName:         "podinfo",
-			ShortDescription:    "Podinfo Helm chart for Kubernetes",
-			SourceUrls:          []string{"https://github.com/stefanprodan/podinfo"},
-			Maintainers: []*corev1.Maintainer{
-				{Name: "stefanprodan", Email: "stefanprodan@users.noreply.github.com"},
-			},
-			Readme:        "Podinfo is used by CNCF projects like [Flux](https://github.com/fluxcd/flux2)",
-			DefaultValues: "Default values for podinfo.\n\nreplicaCount: 1\n",
 		},
 	}
 )
