+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/test_util_test.go
@@ -7,6 +7,7 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"io/ioutil"
 	"net/http"
 	"reflect"
 	"testing"
@@ -143,6 +144,19 @@ func compareJSONStrings(t *testing.T, expectedJSONString, actualJSONString strin
 			t.Errorf("mismatch (-want +got):\n%s", cmp.Diff(expectedJSONString, buf.String()))
 		}
 	}
+}
+
+// generate-cert.sh script in testdata directory is used to generate these files
+func getCertsForTesting(t *testing.T) (ca, pub, priv []byte) {
+	var err error
+	if ca, err = ioutil.ReadFile("testdata/ca.pem"); err != nil {
+		t.Fatalf("%+v", err)
+	} else if pub, err = ioutil.ReadFile("testdata/server.pem"); err != nil {
+		t.Fatalf("%+v", err)
+	} else if priv, err = ioutil.ReadFile("testdata/server-key.pem"); err != nil {
+		t.Fatalf("%+v", err)
+	}
+	return ca, pub, priv
 }
 
 // ref: https://stackoverflow.com/questions/21936332/idiomatic-way-of-requiring-http-basic-auth-in-go
