+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/ca-key.pem
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEICgAqMyEvPEG1bKeX2TcS+sTlgHpWzJteyekDX1zg1/SoAoGCCqGSM49
+AwEHoUQDQgAERcG/6sc0s7XUmT6yfyRVpxTJIq6oKsd9mjqhPfEAsQtZvy5miD4Q
+Dv/AmMOQbBamcDlcj8qs6u83NyrH7jIaMw==
+-----END EC PRIVATE KEY-----
