+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/ca.pem
@@ -1,0 +1,13 @@
+-----BEGIN CERTIFICATE-----
+MIIB4DCCAYWgAwIBAgIUBiiuuxJQ/Elz+YfTWtHzpAX6kLgwCgYIKoZIzj0EAwIw
+RTFDMEEGA1UEAxM6Zmx1eHYycGx1Z2luLXRlc3RkYXRhLXNzbC1zdmMuZGVmYXVs
+dC5zdmMuY2x1c3Rlci5sb2NhbCBDQTAeFw0yMjAzMDIyMzM3MDBaFw0yNzAzMDEy
+MzM3MDBaMEUxQzBBBgNVBAMTOmZsdXh2MnBsdWdpbi10ZXN0ZGF0YS1zc2wtc3Zj
+LmRlZmF1bHQuc3ZjLmNsdXN0ZXIubG9jYWwgQ0EwWTATBgcqhkjOPQIBBggqhkjO
+PQMBBwNCAARFwb/qxzSztdSZPrJ/JFWnFMkirqgqx32aOqE98QCxC1m/LmaIPhAO
+/8CYw5BsFqZwOVyPyqzq7zc3KsfuMhozo1MwUTAOBgNVHQ8BAf8EBAMCAQYwDwYD
+VR0TAQH/BAUwAwEB/zAdBgNVHQ4EFgQU1AF8ckx4/NE0JpgrTGZrRNbmB68wDwYD
+VR0RBAgwBocEfwAAATAKBggqhkjOPQQDAgNJADBGAiEA8Jg71jqePvsZOPLSyUmL
+YBuc3LkpE3egCf5zWdVFl0wCIQCTlLshTuBc3SXX2q0z68bneWvPnp1l0cKCR9Lg
+QwcBGg==
+-----END CERTIFICATE-----
