+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/nginx.conf
@@ -5,12 +5,24 @@ events { }
 
 http {
     server {
-        listen 80;
+        listen 80 default_server;
         root   /usr/share/nginx/html;
 
         location /podinfo-basic-auth/ {
             auth_basic           "Area secured with Basic Authentication";
             auth_basic_user_file /etc/apache2/.htpasswd; 
         }
+
+        # contents of /podinfo-tls can only be accessed over SSL
+        location /podinfo-tls/ {
+            deny all;
+       }
     }
-}    
+
+    server {
+        listen 443 ssl;
+        root   /usr/share/nginx/html/podinfo-tls/;
+        ssl_certificate /etc/ssl/certs/ssl-bundle.pem;
+        ssl_certificate_key /etc/ssl/certs/server-key.pem;
+    }
+}
