+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/podinfo-tls-index.yaml
@@ -1,0 +1,41 @@
+# Copyright 2021-2022 the Kubeapps contributors.
+# SPDX-License-Identifier: Apache-2.0
+
+# This is local copy of the first few entries
+# on "https://stefanprodan.github.io/podinfo/index.yaml" as of Sept 10 2021 with the chart
+# urls modified to link to .tgz files also within the local cluster
+apiVersion: v1
+entries:
+  podinfo:
+    - apiVersion: v1
+      appVersion: 6.0.0
+      created: "2021-06-16T12:51:07.137321429Z"
+      description: Podinfo Helm chart for Kubernetes
+      digest: aed86105891a9b1db588128878b79425cf6447e69b1012119e82c1d37c68ae67
+      home: https://github.com/stefanprodan/podinfo
+      kubeVersion: ">=1.19.0-0"
+      maintainers:
+        - email: stefanprodan@users.noreply.github.com
+          name: stefanprodan
+      name: podinfo
+      sources:
+        - https://github.com/stefanprodan/podinfo
+      urls:
+        - https://fluxv2plugin-testdata-ssl-svc.default.svc.cluster.local:443/podinfo-6.0.0.tgz
+      version: 6.0.0
+    - apiVersion: v1
+      appVersion: 5.2.1
+      created: "2021-05-13T12:57:54.003168505Z"
+      description: Podinfo Helm chart for Kubernetes
+      digest: 6c3cc3b955bce1686036ae6822ee2ca0ef6ecb994e3f2d19eaf3ec03dcba84b3
+      home: https://github.com/stefanprodan/podinfo
+      maintainers:
+        - email: stefanprodan@users.noreply.github.com
+          name: stefanprodan
+      name: podinfo
+      sources:
+        - https://github.com/stefanprodan/podinfo
+      urls:
+        - https://fluxv2plugin-testdata-ssl-svc.default.svc.cluster.local:443/podinfo-5.2.1.tgz
+      version: 5.2.1
+generated: "2021-06-16T12:51:07.135685319Z"
