+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/server-csr.json
@@ -1,0 +1,8 @@
+{
+  "CN": "fluxv2plugin-testdata-ssl-svc.default.svc.cluster.local",
+  "hosts": [
+    "127.0.0.1",
+    "localhost",
+    "fluxv2plugin-testdata-ssl-svc.default.svc.cluster.local"
+  ]
+}
