+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/server-key.pem
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIO1iDVfUeqGKmDdSs6kQqJyaQYbGXaOXe9/niQEOpe0loAoGCCqGSM49
+AwEHoUQDQgAECk2cZkBTcw4JQlaoz5sNDnT7bga4cOMI7t93QNvn9uBYRS7TAUvj
+ni4BONDstTF26yc7B5LsqynSOkgVYJaStA==
+-----END EC PRIVATE KEY-----
