+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/testdata/ssl-bundle.pem
@@ -1,0 +1,28 @@
+-----BEGIN CERTIFICATE-----
+MIICXjCCAgWgAwIBAgIUHYDPnjnIE7gu5W5RDvS40PSoVGswCgYIKoZIzj0EAwIw
+RTFDMEEGA1UEAxM6Zmx1eHYycGx1Z2luLXRlc3RkYXRhLXNzbC1zdmMuZGVmYXVs
+dC5zdmMuY2x1c3Rlci5sb2NhbCBDQTAeFw0yMjAzMDIyMzQyMDBaFw0zMjAyMjgy
+MzQyMDBaMEIxQDA+BgNVBAMTN2ZsdXh2MnBsdWdpbi10ZXN0ZGF0YS1zc2wtc3Zj
+LmRlZmF1bHQuc3ZjLmNsdXN0ZXIubG9jYWwwWTATBgcqhkjOPQIBBggqhkjOPQMB
+BwNCAAQKTZxmQFNzDglCVqjPmw0OdPtuBrhw4wju33dA2+f24FhFLtMBS+OeLgE4
+0Oy1MXbrJzsHkuyrKdI6SBVglpK0o4HVMIHSMA4GA1UdDwEB/wQEAwIFoDAdBgNV
+HSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwDAYDVR0TAQH/BAIwADAdBgNVHQ4E
+FgQUKyEA6PY/qMRrf1HW6tx3jskxkvYwHwYDVR0jBBgwFoAU1AF8ckx4/NE0Jpgr
+TGZrRNbmB68wUwYDVR0RBEwwSoIJbG9jYWxob3N0gjdmbHV4djJwbHVnaW4tdGVz
+dGRhdGEtc3NsLXN2Yy5kZWZhdWx0LnN2Yy5jbHVzdGVyLmxvY2FshwR/AAABMAoG
+CCqGSM49BAMCA0cAMEQCIBTZjKw8G1/OiCj2wkt/0uZkJqkJ3FWdW8Np8Hul5od9
+AiAIR5dSzWB36J6GvYqtL5vnIx4OMpQpwX3haY3wcRaIsQ==
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIB4DCCAYWgAwIBAgIUBiiuuxJQ/Elz+YfTWtHzpAX6kLgwCgYIKoZIzj0EAwIw
+RTFDMEEGA1UEAxM6Zmx1eHYycGx1Z2luLXRlc3RkYXRhLXNzbC1zdmMuZGVmYXVs
+dC5zdmMuY2x1c3Rlci5sb2NhbCBDQTAeFw0yMjAzMDIyMzM3MDBaFw0yNzAzMDEy
+MzM3MDBaMEUxQzBBBgNVBAMTOmZsdXh2MnBsdWdpbi10ZXN0ZGF0YS1zc2wtc3Zj
+LmRlZmF1bHQuc3ZjLmNsdXN0ZXIubG9jYWwgQ0EwWTATBgcqhkjOPQIBBggqhkjO
+PQMBBwNCAARFwb/qxzSztdSZPrJ/JFWnFMkirqgqx32aOqE98QCxC1m/LmaIPhAO
+/8CYw5BsFqZwOVyPyqzq7zc3KsfuMhozo1MwUTAOBgNVHQ8BAf8EBAMCAQYwDwYD
+VR0TAQH/BAUwAwEB/zAdBgNVHQ4EFgQU1AF8ckx4/NE0JpgrTGZrRNbmB68wDwYD
+VR0RBAgwBocEfwAAATAKBggqhkjOPQQDAgNJADBGAiEA8Jg71jqePvsZOPLSyUmL
+YBuc3LkpE3egCf5zWdVFl0wCIQCTlLshTuBc3SXX2q0z68bneWvPnp1l0cKCR9Lg
+QwcBGg==
+-----END CERTIFICATE-----
